画像とスタイルの変更を削除しました
