+++ src/assets/scss/components/_index.scss
@@ -1,1 +1,1 @@
-@forward "./component";
+@forward "./_drawer-button";
