+++ src/assets/scss/config/_base.scss
@@ -10,7 +10,7 @@
 
   --inner: calc(var(--vw) - var(--vw) / 24 * 2);
 
-  --col: calc((var(--vw)) / 24);
+  --col: calc((var(--vw)) / 28);
 
   --margin: 4rem;
 
@@ -156,6 +156,27 @@ ol {
   }
 }
 
+._en {
+  font-family:
+    "Jost",
+    system-ui,
+    -apple-system,
+    BlinkMacSystemFont,
+    "Segoe UI",
+    Roboto,
+    Oxygen,
+    Ubuntu,
+    Cantarell,
+    "Open Sans",
+    "Helvetica Neue",
+    sans-serif;
+}
+
+._serif {
+  font-family: "Noto Serif JP", "Times New Roman", "YuMincho", "Hiragino Mincho ProN", "Yu Mincho", "MS PMincho",
+    system-ui, serif;
+}
+
 //font-size
 :root {
   --font-size-small-3: calc(1.5rem * 8 / 11);
@@ -172,19 +193,19 @@ ol {
   --font-size-8: calc(1.5rem * 8 / 0.8);
   --font-size-9: calc(1.5rem * 8 / 0.6);
   @include mq(sp) {
-    --font-size-small-3: calc(1.3rem * 8 / 11);
-    --font-size-small-2: calc(1.3rem * 8 / 10);
-    --font-size-small-1: calc(1.3rem * 8 / 9);
-    --font-size-base: calc(1.3rem * 8 / 8);
-    --font-size-1: calc(1.3rem * 8 / 7);
-    --font-size-2: calc(1.3rem * 8 / 6);
-    --font-size-3: calc(1.3rem * 8 / 5);
-    --font-size-4: calc(1.3rem * 8 / 4);
-    --font-size-5: calc(1.3rem * 8 / 3);
-    --font-size-6: calc(1.3rem * 8 / 2);
-    --font-size-7: calc(1.3rem * 8 / 1);
-    --font-size-8: calc(1.3rem * 8 / 0.8);
-    --font-size-9: calc(1.3rem * 8 / 0.6);
+    --font-size-small-3: calc(1.3rem * 16 / 19);
+    --font-size-small-2: calc(1.3rem * 16 / 18);
+    --font-size-small-1: calc(1.3rem * 16 / 17);
+    --font-size-base: calc(1.3rem * 16 / 16);
+    --font-size-1: calc(1.3rem * 16 / 15);
+    --font-size-2: calc(1.3rem * 16 / 14);
+    --font-size-3: calc(1.3rem * 16 / 13);
+    --font-size-4: calc(1.3rem * 16 / 12);
+    --font-size-5: calc(1.3rem * 16 / 11);
+    --font-size-6: calc(1.3rem * 16 / 10);
+    --font-size-7: calc(1.3rem * 16 / 9);
+    --font-size-8: calc(1.3rem * 16 / 8);
+    --font-size-9: calc(1.3rem * 16 / 7);
   }
 }
