+++ src/assets/scss/config/_fonts.scss
@@ -64,23 +64,3 @@ html {
   }
 }
 
-._en {
-  font-family:
-    "Jost",
-    system-ui,
-    -apple-system,
-    BlinkMacSystemFont,
-    "Segoe UI",
-    Roboto,
-    Oxygen,
-    Ubuntu,
-    Cantarell,
-    "Open Sans",
-    "Helvetica Neue",
-    sans-serif;
-}
-
-._serif {
-  font-family: "Noto Serif JP", "Times New Roman", "YuMincho", "Hiragino Mincho ProN", "Yu Mincho", "MS PMincho",
-    system-ui, serif;
-}
