--- src/assets/scss/pages/home/_index.scss
@@ -1,1 +1,0 @@
-// @forward "./";
