+++ src/assets/scss/pages/home/index.scss
@@ -1,0 +1,1 @@
+@forward "./_home-demo";
