+++ src/assets/scss/projects/index.scss
@@ -1,1 +1,1 @@
-@forward "./project";
+@forward "./_list-drawer-nav";
