+++ src/includes/components/drawer-button.html
@@ -1,4 +1,14 @@
-<button class="drawer-button">
-  <div class="drawer-button__text--menu">menu</div>
-  <div class="drawer-button__text--close">close</div>
-</button>
+<div class="drawer-button">
+  <div class="drawer-button__inner">
+    <div class="drawer-button__row">
+      <button class="drawer-button__text drawer-button__text--menu _en">
+        menu
+      </button>
+    </div>
+    <div class="drawer-button__row">
+      <button class="drawer-button__text drawer-button__text--close _en">
+        close
+      </button>
+    </div>
+  </div>
+</div>
