+++ src/includes/globals/drawer-nav.html
@@ -1,13 +1,25 @@
 <nav class="drawer-nav">
   <div class="drawer-nav__inner">
-    <ul class="drawer-nav__list">
-      <li class="drawer-nav__list__item">
-        <a class="drawer-nav__list__item__link" href="/"> home </a>
+    <ul class="list-drawer-nav">
+      <li class="item-drawer-nav">
+        <a
+          class="item-drawer-nav__link _en"
+          href="https://twitter.com/webcreaterfrm30"
+          target="_blank"
+        >
+          X
+        </a>
       </li>
-      <li class="drawer-nav__list__item">
-        <a class="drawer-nav__list__item__link" href="/"> sns </a>
+      <li class="item-drawer-nav">
+        <a
+          class="item-drawer-nav__link _en"
+          href="https://github.com/yukaorange"
+          target="_blank"
+        >
+          github
+        </a>
       </li>
     </ul>
-    <button class="drawer-nav__close">close</button>
+    <button class="drawer-nav__close _en">close</button>
   </div>
 </nav>
