+++ src/includes/globals/header.html
@@ -1,10 +1,10 @@
 <header class="header">
   <nav class="navigation">
-    <ul class="navigation__list">
+    <!-- <ul class="navigation__list">
       <li class="navigation__list__item">
         <a href="/" class="navigation__list__item__link"> home </a>
       </li>
-    </ul>
+    </ul> -->
   </nav>
 </header>
 <!-- drawer-nav  -->
