+++ src/pages/index.html
@@ -10,9 +10,8 @@
     {{>header}}
     <!-- content  -->
     <div class="content home" data-barba="container" data-template="home">
-      <div class="home">
-        <h2>Home page</h2>
-        <p class="home__text">home page</p>
+      <div class="home-demo">
+        <h2>demo</h2>
       </div>
     </div>
     <div id="webgl"></div>
